fix(models): float to decimal & add decimal places

+++ order/ordering/app/models/order.py
@@ -1,6 +1,7 @@
 from sqlmodel import SQLModel, Field
 from datetime import datetime
 from enum import Enum
+from decimal import Decimal
 
 
 class TypeOrder(str, Enum):
@@ -31,7 +32,7 @@ class Order(SQLModel, table=True):
 
     status: StatusOrder = Field(default=StatusOrder.ORDERED, index=True)
 
-    total_amount: float = Field(default=0.0, index=True)
+    total_amount: Decimal = Field(default=0.0, index=True, decimal_places=2)
 
 
 class OrderDetails(SQLModel, table=True):
@@ -41,7 +42,9 @@ class OrderDetails(SQLModel, table=True):
     product_id: int = Field(foreign_key="product.product_id")
 
     quantity: int = Field(gt=0)
-    line_total: float = Field(gt=0.0)
+
+    # The total of a line of product = quantity * product. price
+    line_total: Decimal = Field(gt=0.0, decimal_places=2)
 
     # TODO: add feature to add TypeOrder
     # type: TypeOrder = Field(default=TypeOrder.EAT_IN, index=True)
